move silx.gui.plot.ImageStack.UrlList to silx.gui.widgets.UrlList
