+++ src/silx/gui/plot/ImageStack.py
@@ -33,6 +33,7 @@ from silx.gui.plot import Plot2D
 from silx.io.url import DataUrl
 from silx.io.utils import get_data
 from silx.gui.widgets.FrameBrowser import HorizontalSliderWithBrowser
+from silx.gui.widgets.UrlList import UrlList
 from silx.gui.utils import blockSignals
 
 import typing
@@ -57,110 +58,6 @@ class _HorizontalSlider(HorizontalSliderWithBrowser):
 
     def _urlChanged(self, value):
         self.sigCurrentUrlIndexChanged.emit(value)
-
-
-class UrlList(qt.QListWidget):
-    """List of URLs the user to select an URL"""
-
-    sigCurrentUrlChanged = qt.Signal(str)
-    """Signal emitted when the active/current URL has changed.
-
-    This signal emits the empty string when there is no longer an active URL.
-    """
-
-    sigUrlRemoved = qt.Signal(str)
-    """Signal emit when an url is removed from the URL list.
-
-    Provides the url (DataUrl) as a string
-    """
-
-    def __init__(self, parent=None):
-        super().__init__(parent)
-        self._editable = False
-        # are we in 'editable' mode: for now if true then we can remove some items from the list
-
-        # menu to be triggered when in edition from right-click
-        self._menu = qt.QMenu()
-        self._removeAction = qt.QAction(
-            text="Remove",
-            parent=self
-        )
-        self._removeAction.setShortcuts(
-            [
-                # qt.Qt.Key_Delete,
-                qt.QKeySequence.Delete,
-            ]
-        )
-        self._menu.addAction(self._removeAction)
-
-        # connect signal / Slot
-        self.currentItemChanged.connect(self._notifyCurrentUrlChanged)
-
-    def setEditable(self, editable: bool):
-        """Toggle whether the user can remove some URLs from the list"""
-        if editable != self._editable:
-            self._editable = editable
-            # discusable choice: should we change the selection mode ? No much meaning
-            # to be in ExtendedSelection if we are not in editable mode. But does it has more
-            # meaning to change the selection mode ?
-            if editable:
-                self._removeAction.triggered.connect(self._removeSelectedItems)
-                self.addAction(self._removeAction)
-            else:
-                self._removeAction.triggered.disconnect(self._removeSelectedItems)
-                self.removeAction(self._removeAction)
-
-    def setUrls(self, urls: list) -> None:
-        url_names = []
-        [url_names.append(url.path()) for url in urls]
-        self.addItems(url_names)
-
-    def removeUrl(self, url: str):
-        sel_items = self.findItems(url, qt.Qt.MatchExactly)
-        if len(sel_items) > 0:
-            assert len(sel_items) == 0, "at most one item expected"
-            self.removeItemWidget(sel_items[0])
-
-    def _notifyCurrentUrlChanged(self, current, previous):
-        if current is None:
-            self.sigCurrentUrlChanged.emit("")
-        else:
-            self.sigCurrentUrlChanged.emit(current.text())
-
-    def setUrl(self, url: typing.Optional[DataUrl]) -> None:
-        """Set the current URL.
-
-        :param url: The new selected URL. Use `None` to clear the selection.
-        """
-        if url is None:
-            self.clearSelection()
-            self.sigCurrentUrlChanged.emit("")
-        else:
-            assert isinstance(url, DataUrl)
-            sel_items = self.findItems(url.path(), qt.Qt.MatchExactly)
-            if sel_items is None:
-                _logger.warning(url.path(), ' is not registered in the list.')
-            elif len(sel_items) > 0:
-                item = sel_items[0]
-                self.setCurrentItem(item)
-                self.sigCurrentUrlChanged.emit(item.text())
-
-    def _removeSelectedItems(self):
-        if not self._editable:
-            raise ValueError("UrlList is not set as 'editable'")
-        urls = []
-        for item in self.selectedItems():
-            url = item.text()
-            self.takeItem(self.row(item))
-            urls.append(url)
-        # as the connected slot of 'sigUrlRemoved' can modify the items, better handling all at the end
-        for url in urls:
-            self.sigUrlRemoved.emit(url)
-
-    def contextMenuEvent(self, event):
-        if self._editable:
-            globalPos = self.mapToGlobal(event.pos())
-            self._menu.exec_(globalPos)
 
 
 class _ToggleableUrlSelectionTable(qt.QWidget):
